requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests>=2.32.2
openai>=1.50.0
python-dotenv>=1.0.1 